BUG: Add workaround for jQuery XSS vulnerability
Dependabot gave us a warning about the version of jQuery in use:
apparently you could trigger arbitrary code injection in this version
of jQuery in a silly way. **This should not actually impact Qurro**,
because the visualization side of Qurro is only run locally in the
browser -- I believe the only way you could trigger it would be messing
with the data you provide to Qurro, and at that point it's not really
our problem.

It's analogous to how a bash script that e.g. echoes a user-provided
input is in theory vulnerable to shell injection, I think -- the user of
Qurro, and/or the person who provides them the data, would have to be
acting maliciously. And even then the impacts of this should be very
limited.

In any case, this implements the suggested workaround so that we
can remove this warning.